Refactor: 사용하지 않는 export 제거 및 파일 위치 수정
