+++ src/pages/admin/AdminSetStatePage.tsx
@@ -5,9 +5,9 @@ import CommonBtn from '@gdsc/components/common/button/CommonBtn';
 import { DisplayLayout } from '@gdsc/styles/LayoutStyle';
 
 import { ButtonContainer, TopContainer } from './AdminSetStatePage.style';
-import AdminSearchBar from './components/AdminSearchBar';
-import DeleteModal from './components/DeleteModal';
-import MemberTable from './components/MemberTable';
+import AdminSearchBar from './components/status/AdminSearchBar';
+import DeleteModal from './components/status/DeleteModal';
+import MemberTable from './components/status/MemberTable';
 
 const AdminSetStatePage = () => {
   const [deleteModalOpen, setDeleteModalOpen] = useState<boolean>(false);
